# OPA: Allow prepare query optimisation via partial evaluation