This adds a command line flag that enables an optimisation of OPA policies before they are evaluated in the respective filters. This preparation optimises the policies for cpu at the expense of using more memory and uses a core OPA feature.  

Missing:
- [ ] Wait for upstream [PR](https://github.com/open-policy-agent/opa-envoy-plugin/pull/604) to merged
- [ ] Benchmarks on a medium complex policy (cpu vs memory)