@Pushpalanka 2024-12-08
Benchmark results on commit https://github.com/zalando/skipper/commit/02588ce096089299d1698ef051cbb89a59a3df75

**Observation**:  
For the exact pure JWT validation policy could see a slight improvement with pre-evaluation against without it. 
But this policy in current commit with additional data.json involved rules, optimization was negatively affected for CPU and memory performance (if the flag usage correctly understood).

Carried out with `authorize-request-jwt-validation-with-pre-evaluation` benchmark test, setting the `pre-evaluation` flag true vs false.
goos: darwin
goarch: arm64
cpu: Apple M1

### With pre-evaluation
 -memprofile pre-eval-true-mem5.prof
195                         &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; 5868242 ns/op     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;                   2657280 B/op      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;             40073 allocs/op
3.592s

 -cpuprofile pre-eval-true5.prof
199     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;      5845913 ns/op    &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     2654548 B/op  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    40012 allocs/op
3.302s

### With pre-evaluation flag false

 -memprofile pre-eval-false-mem5.prof
9117     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;        125310 ns/op     &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;       77774 B/op   &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;     1299 allocs/op
3.090s

 -cpuprofile pre-eval-false5.prof
9334      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;       121684 ns/op    &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;        77735 B/op    &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    1299 allocs/op
3.054s

### Memory
Without pre-evaluation
<img width="1434" alt="image" src="https://github.com/user-attachments/assets/92e365cf-4927-4a5a-8452-cec60b66dd7b">

With pre-evaluation
<img width="1434" alt="image" src="https://github.com/user-attachments/assets/4a0dfaea-6f34-420e-80c2-76c7d3609ed7">

### CPU
Without pre evaluation
<img width="1434" alt="image" src="https://github.com/user-attachments/assets/c308198b-0722-4761-a67a-f9381a2db38e">

With pre-evaluation
<img width="1434" alt="image" src="https://github.com/user-attachments/assets/bf069180-8281-41ed-b2a5-3801884504ed">

Profiles attached.
[pre-eval-false-cpu.txt](https://github.com/user-attachments/files/18054210/pre-eval-false-cpu.txt)
[pre-eval-true-cpu.txt](https://github.com/user-attachments/files/18054211/pre-eval-true-cpu.txt)
[pre-eval-true-mem.txt](https://github.com/user-attachments/files/18054212/pre-eval-true-mem.txt)
[pre-eval-false-mem.txt](https://github.com/user-attachments/files/18054213/pre-eval-false-mem.txt)

May be a test using a much larger data source via a bundle would provide more insights.